% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_angles.R
\name{polyg_angles}
\alias{polyg_angles}
\title{Minimum and maximum angles between two \code{sf} LINESTRING objects}
\usage{
polyg_angles(input1, input2)
}
\arguments{
\item{input1}{object of class sf or sfc of type LINESTRING or MULTILINESTRING.}

\item{input2}{object of class sf or sfc of type LINESTRING or MULTILINESTRING.}
}
\value{
An sf object equivalent to \code{input1} with added columns containing
minimum and maximum angles.
}
\description{
Function for computing the minimum and maximum angles
between two intersecting \code{sf} MULTILINESTRING or LINESTRING objects.
}
\details{
Details
}
