% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_contiguity.R
\name{polyg_contiguity}
\alias{polyg_contiguity}
\title{Common contiguities of an \code{sf} POLYGON object}
\usage{
polyg_contiguity(polygons_input, contiguity = "queen", fill_empty = FALSE)
}
\arguments{
\item{polygons_input}{object of class sf, sfc or sfg with geometry type POLYGON or MULTIPOLYGON.}

\item{contiguity}{character; either \code{"queen"}, \code{"rook"} or \code{"bishop"}. Default is \code{"queen"}.}

\item{fill_empty}{logical; should empty indices be filled? Default is \code{FALSE}.}
}
\value{
A sparse index list of contiguities.
}
\description{
Wrapper function for \code{\link[sf:st_relate]{sf::st_relate()}}. Computes a sparse index
list of contiguities for an \code{sf} POLYGON object.
}
\details{
Wrapper for \code{\link[sf:st_relate]{sf::st_relate()}} with DE9-IM predicate \verb{F***T****} for queen
contiguity, \verb{F***1****} for rook contiguity and \verb{F***0****} for bishop contiguity.
Setting \code{fill_empty} to \code{TRUE} guarantees at least one index per polygon by adding self-neighboring.
}
