% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_dissolve_enclosed.R
\name{polyg_dissolve_enclosed}
\alias{polyg_dissolve_enclosed}
\title{Dissolve enclosed polygons in a sf POLYGON object}
\usage{
polyg_dissolve_enclosed(
  polygons_input,
  contiguity = "rook",
  tolerance = 0,
  verbose = FALSE
)
}
\arguments{
\item{polygons_input}{object of class sf, sfc or sfg with geometry type POLYGON or MULTIPOLYGON.}

\item{contiguity}{character; either \code{"rook"} or \code{"queen"}. Default is \code{"rook"}.}

\item{tolerance}{numeric; length in units to tolerate when determining if a polygon is enclosed.
Default is 0 (i.e. fully enclosed).}

\item{verbose}{logical; should number of dissolved polygons be printed? Default is \code{FALSE}.}
}
\value{
An sf object of type POLYGON with enclosed polygons dissolved.
}
\description{
Helper function for dissolving fully or partially enclosed polygons in a \code{sf} POLYGON object.
}
\details{
An enclosed polygon is defined as having only one neighbor which shares all of the polygon border.
Setting tolerance > 0 allows for merging partially enclosed polygons, i.e. polygons where small portions
of the polygon border is not shared with any other polygon.
}
