% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_leaf_nodes.R
\name{polyg_leaf_nodes}
\alias{polyg_leaf_nodes}
\title{Leaf nodes of an sf LINESTRING object}
\usage{
polyg_leaf_nodes(input)
}
\arguments{
\item{input}{object of class sf, sfc or sfg with geometry type LINESTRING or MULTILINESTRING.}
}
\value{
An sf object of type POINT containing all linestring leaf nodes.
}
\description{
Helper function for computing leaf nodes of an sf LINESTRING object.
}
\details{
Given a linestring object we can determine \code{A} = \emph{linestring intersections} and \code{B} = \emph{linestring endpoints}.
The function then computes \verb{A\\B}, i.e. the relative complement of \code{A} in \code{B}. If the input contains geometries of type
MULTILINESTRING, they are cast to LINESTRING before computing.
}
