% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_rename_geom.R
\name{polyg_rename_geom}
\alias{polyg_rename_geom}
\title{Rename the geometry of an \code{sf} object}
\usage{
polyg_rename_geom(input, geom_name = "geometry")
}
\arguments{
\item{input}{object of class sf}

\item{geom_name}{desired name for the geometry column. Defaults to "geometry"}
}
\value{
The input sf object with renamed geometry column
}
\description{
Simple renaming function for the geometry column of sf objects
}
