% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_contiguity.R
\name{st_contiguity}
\alias{st_contiguity}
\title{Common contiguities of an sf POLYGON object}
\usage{
st_contiguity(x, contiguity = "queen", fill_empty = FALSE)
}
\arguments{
\item{x}{object of class sf, sfc or sfg with geometry type POLYGON or MULTIPOLYGON.}

\item{contiguity}{character vector of length 1, which is wither "queen", "rook" or "bishop". Defaults to \code{"queen"}.}

\item{fill_empty}{logical; should empty indices be filled? Defaults to \code{FALSE}.}
}
\value{
A sparse index list of contiguities.
}
\description{
Wrapper function for \code{\link[sf:st_relate]{sf::st_relate()}}. Computes a sparse index list of contiguities for an sf POLYGON object.
}
\details{
This function is a wrapper for \code{\link[sf:st_relate]{sf::st_relate()}} with DE9-IM predicate \verb{F***T****} for queen
contiguity, \verb{F***1****} for rook contiguity and \verb{F***0****} for bishop contiguity.
Setting \code{fill_empty} to \code{TRUE} guarantees at least one index per polygon by adding self-neighboring.
}
