% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_deadends.R
\name{st_deadends}
\alias{st_deadends}
\title{Compute dead ends of an sf linestring object}
\usage{
st_deadends(x)
}
\arguments{
\item{x}{object of class sf, sfc or sfg with geometry type LINESTRING or MULTILINESTRING.}
}
\value{
An sf object of type POINT containing all linestring dead ends.
}
\description{
Compute dead ends of an sf linestring object.
}
\details{
Given a linestring object where \code{A} = \emph{intersections} and \code{B} = \emph{endpoints}, the function
computes \verb{A \\ B}, i.e. the relative complement of \code{A} in \code{B}. If the input contains geometries of type
MULTILINESTRING, they are cast to LINESTRING before computing.
}
