% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nearest_node.R
\name{st_nearest_node}
\alias{st_nearest_node}
\title{Compute a linestring connection from an sf POINT object to nodes of a sf LINESTRING object}
\usage{
st_nearest_node(points, lines, parallel = FALSE)
}
\arguments{
\item{points}{object of class sf, sfc or sfg with geometry type POINT or MULTIPOINT.}

\item{lines}{object of class sf, sfc or sfg with geometry type LINESTRING or MULTILINESTRING.}

\item{parallel}{logical; should the computation be in parallel? Defaults to FALSE.}
}
\value{
An sf object of type LINESTRING containing the minimum distance linestring
connection from the \code{point} input to the nodes of the \code{lines} input.
}
\description{
Compute the minimum distance linestring connection from an sf POINT object to nodes of an sf LINESTRING object
}
\details{
Given a point, the function computes the minimum distance tocthe nodes of a set of
linestrings and outputs the relevant linestring. If the point overlaps
with a node, this node is not considered in the computation.
This guarantees linestrings to have lengh > 0.
}
