% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rook.R
\name{st_rook}
\alias{st_rook}
\title{Compute rook contiguity of an sf object}
\usage{
st_rook(x, fill_empty = FALSE)
}
\arguments{
\item{x}{object of class sf, sfc or sfg with geometry type POLYGON or MULTIPOLYGON.}

\item{fill_empty}{logical; should empty indices be filled (\code{TRUE})?}
}
\value{
A sparse index list of rook contiguities.
}
\description{
Wrapper function for \code{\link[sf:st_relate]{sf::st_relate()}}. Computes a sparse index list of rook contiguities for an sf polygon object.
}
\details{
This function is a wrapper for \code{\link[sf:st_relate]{sf::st_relate()}} with DE9-IM pattern \verb{F***1****}.
Setting \code{fill_empty} to \code{TRUE} guarantees at least one index per polygon by adding self-neighboring.
}
