% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_cast_substring.R
\name{polyg_cast_substring}
\alias{polyg_cast_substring}
\title{Cast an \code{sf} object to substring}
\usage{
polyg_cast_substring(input_lines)
}
\arguments{
\item{input_lines}{object of class sf or sfc with geometry type MULTILINESTRING or LINESTRING.}
}
\value{
An sf or sfc object of type LINESTRING containing the substrings of \code{input_lines}.
}
\description{
Helper function for computing the substrings of an \code{sf} MULTILINESTRING or LINESTRING object.
}
\details{
The output is always LINESTRING of class matching the \code{input_lines} class.
}
