% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_leaf_nodes.R
\name{polyg_leaf_nodes}
\alias{polyg_leaf_nodes}
\title{Identify leaf nodes of an \code{sf} LINESTRING object}
\usage{
polyg_leaf_nodes(input_lines)
}
\arguments{
\item{input_lines}{object of class sf of type LINESTRING or MULTILINESTRING.}
}
\value{
An sf object of type POINT containing all leaf nodes of \code{input}.
}
\description{
Helper function for computing leaf nodes of an \code{sf} MULTILINESTRING or LINESTRING object.
}
\details{
Function to compute leaf nodes, i.e. nodes which only intersect
with one linestring in the \code{sf} LINESTRING object. If \code{input} contains geometries
of type MULTILINESTRING, they are cast to LINESTRING before computing.
}
