% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyg_nearest_node.R
\name{polyg_nearest_node}
\alias{polyg_nearest_node}
\title{Shortest linestring connections from an sf POINT object to nodes of a sf LINESTRING object}
\usage{
polyg_nearest_node(points, lines, parallel = FALSE)
}
\arguments{
\item{points}{object of class sf, sfc or sfg with geometry type POINT or MULTIPOINT.}

\item{lines}{object of class sf, sfc or sfg with geometry type LINESTRING or MULTILINESTRING.}

\item{parallel}{logical; should the computation be in parallel? Defaults to FALSE.}
}
\value{
An sf object of type LINESTRING containing the minimum distance linestring
connection from the \code{points} input to nodes of the \code{lines} input.
}
\description{
Compute the minimum distance linestring connection from an sf POINT object to nodes of an sf LINESTRING object
}
\details{
Given a point, the function computes the minimum distance to the nodes of a set of
linestrings and outputs the relevant linestring. If the point already overlaps
with an existing node, this node is disregarded.
This guarantees linestrings to have length > 0. If the \code{points} or \code{lines} input contains
geometries of type MULTIPOINT or MULTILINESTRING, they are converted to POINT and LINESTRING before computing.
}
